completed parser tests, moved to seperate file
