+++ neurovault/apps/statmaps/tests/test_nidm.py
@@ -1,12 +1,9 @@
-from django.test import TestCase, Client, override_settings
-from django.core.urlresolvers import reverse
+from django.test import TestCase, Client
 from neurovault.apps.statmaps.models import Collection,User
-from uuid import uuid4
 import tempfile
 import os
 import shutil
-import zipfile
-from neurovault.apps.statmaps.utils import NIDMUpload
+from neurovault.apps.statmaps.nidm_results import NIDMUpload
 
 
 class NIDMResultsTest(TestCase):
@@ -14,13 +11,32 @@ class NIDMResultsTest(TestCase):
     def setUp(self):
         testpath = os.path.abspath(os.path.dirname(__file__))
         self.files = {
-            'fsl_nidm':      os.path.join(testpath,'test_data/nidm/fsl.nidm.zip'),
-            'two_contrasts': os.path.join(testpath,'test_data/nidm/two_contrasts.nidm.zip'),
+            'fsl_nidm': {
+                'file': os.path.join(testpath,'test_data/nidm/fsl.nidm.zip'),
+                'output_row': {'type': u'TStatistic','name': 'Generation TStatistic',},
+                'num_statmaps': 2,
+            },
+            'two_contrasts': {
+                'file': os.path.join(testpath,'test_data/nidm/two_contrasts.nidm.zip'),
+                'output_row': {'type': u'FStatistic', 'name': 'Generation FStatistic',},
+                'num_statmaps': 18,
+            },
+            'spm_example': {
+                'file': os.path.join(testpath,'test_data/nidm/spm_example.nidm.zip'),
+                'output_row': {'type': u'TStatistic', 'name': u'passive listening > rest', },
+                'num_statmaps': 1,
+            },
+            # case for a zip with no enclosed directory
+            'spm_nosubdir': {
+                'file': os.path.join(testpath,'test_data/nidm/spm_nosubdir.nidm.zip'),
+                'output_row': {'type': u'FStatistic', 'name': u'Generation FStatistic',},
+                'num_statmaps': 18,
+            },
+        }
 
-            ## not parsing with latest rdflib 4.1.2
-            ## turtle parsing still under active development
-            ## see also: https://github.com/RDFLib/rdflib/issues/336
-            'broken_spm_example':   os.path.join(testpath,'test_data/nidm/spm_example.nidm.zip'),
+        self.failing_files = {
+            #corrupted ttl
+            'spm_bad_ttl':   os.path.join(testpath,'test_data/nidm/spm_bad_ttl.nidm.zip'),
         }
 
         self.tmpdir = tempfile.mkdtemp()
@@ -40,9 +56,9 @@ class NIDMResultsTest(TestCase):
         uploads = {}
         statmaps = {}
 
-        for name in ['fsl_nidm','two_contrasts']:
-            uploads[name] = NIDMUpload(self.files[name],tmp_dir=self.tmpdir,load=False)
-            turtles[name] = uploads[name].parse_ttl(extract=True)
+        for name in self.files:
+            uploads[name] = NIDMUpload(self.files[name]['file'],tmp_dir=self.tmpdir,load=False)
+            turtles[name] = uploads[name].parse_metafiles(extract_ttl=True)
             contrasts[name] = uploads[name].parse_contrasts()
             statmaps[name] = uploads[name].get_statmaps()
 
@@ -50,4 +66,14 @@ class NIDMResultsTest(TestCase):
         assert known bad file throws parsing error
         """
         with self.assertRaises(NIDMUpload.ParseException):
-            bad_ttl = NIDMUpload(self.files['broken_spm_example'])
+            print NIDMUpload(self.failing_files['spm_bad_ttl'])
+
+        """
+        assert output matches expected output
+        """
+        for name, info in self.files.items():
+            first_map = sorted(statmaps[name])[0]
+            for field in 'name','type':
+                self.assertEquals(first_map[field],info['output_row'][field])
+            self.assertEquals(len(statmaps[name]), info['num_statmaps'])
+
