+++ neurovault/apps/statmaps/utils.py
@@ -19,12 +19,6 @@ import nibabel as nib
 from django.core.files.base import ContentFile
 from django.core.files.uploadedfile import InMemoryUploadedFile
 from ast import literal_eval
-import zipfile
-from fnmatch import fnmatch
-import rdflib
-from rdflib.plugins.parsers.notation3 import BadSyntax
-from collections import Counter
-from urlparse import urlparse
 
 
 # see CollectionRedirectMiddleware
@@ -274,145 +268,4 @@ def memory_uploadfile(new_file, fname, old_file):
 
     return InMemoryUploadedFile(cfile, "file", fname,
                                 content_type, cfile.size, charset)
-
-
-class NIDMUpload:
-
-    def __init__(self, zip_path,tmp_dir=None,load=True):
-        self.path = zip_path
-        self.ttl_relpath = ''
-        self.workdir = os.path.split(zip_path)[0] if tmp_dir is None else tmp_dir
-        self.zipobj = None
-        self.raw_ttl = ''
-        self.valid_ttl = False
-        self.unzipped = False
-        self.contrasts = []
-        self.statmaps = []
-
-        if load:
-            self.get_statmaps()
-
-    class ParseException(Exception):
-        pass
-
-    class NoStatMapsException(Exception):
-        pass
-
-    def parse_ttl(self,extract=False):
-        try:
-            self.zipobj = zipfile.ZipFile(self.path)
-        except Exception:
-            raise self.ParseException("Unable to read the zip file.")
-        ttl_list = [v for v in self.zipobj.infolist() if fnmatch(v.filename,'*.ttl')]
-        if len(ttl_list) > 1:
-            raise self.ParseException("Detected more than one .ttl file in zip.")
-        if not ttl_list:
-            return False
-
-        self.raw_ttl = self.zipobj.read(ttl_list[0])
-        self.ttl_relpath = os.path.dirname(ttl_list[0].filename)
-        return self.raw_ttl if extract else True
-
-    def parse_contrasts(self):
-        query = """
-        prefix prov: <http://www.w3.org/ns/prov#>
-        prefix nidm: <http://www.incf.org/ns/nidash/nidm#>
-        prefix spm: <http://www.incf.org/ns/nidash/spm#>
-        prefix fsl: <http://www.incf.org/ns/nidash/fsl#>
-
-        SELECT ?contrastName ?statFile ?statType WHERE {
-         ?cid a nidm:ContrastMap ;
-              nidm:contrastName ?contrastName ;
-              prov:atLocation ?cfile .
-         ?cea a nidm:ContrastEstimation .
-         ?cid prov:wasGeneratedBy ?cea .
-         ?sid a nidm:StatisticMap ;
-              nidm:statisticType ?statType ;
-              prov:atLocation ?statFile .
-        }
-        """
-        nidm_g = rdflib.Graph()
-
-        try:
-            nidm_g.parse(data=self.raw_ttl, format='turtle')
-            self.valid_ttl = True
-        except BadSyntax:
-            raise self.ParseException("RDFLib was unable to parse the .ttl file.")
-
-        c_results = nidm_g.query(query)
-        for row in c_results.bindings:
-            c_row = {}
-            for key, val in sorted(row.items()):
-                c_row[str(key)] = val.decode()
-            self.contrasts.append(c_row)
-        return self.contrasts
-
-    def unpack_nidm_zip(self):
-        if not self.raw_ttl:
-            self.parse_ttl()
-        if not self.contrasts:
-            self.parse_contrasts()
-
-        try:
-            self.zipobj.extractall(path=self.workdir)
-            self.unzipped = True
-        except Exception,e:
-            raise self.ParseException("Unable to unzip: %s" % e)
-
-    def get_statmaps(self):
-        if not self.unzipped:
-            self.unpack_nidm_zip()
-        if not self.contrasts:
-            raise self.NoStatMapsException("No eligible data found in this file")
-
-        incidences = Counter([v['contrastName'] for v in self.contrasts])
-        for contrast in self.contrasts:
-            map = {'name':contrast['contrastName'],
-                   'type':self.parse_statmap_type(contrast['statType']),
-                   'file':self.validate_statmap_uri(contrast)}
-
-            if incidences[contrast['contrastName']] > 1:
-                map['name'] = self.get_unique_mapname(contrast)
-            self.statmaps.append(map)
-        return self.statmaps
-
-    def validate_statmap_uri(self,contrast):
-        rel_path = self.uri_to_path(contrast['statFile'])
-        path = self.expand_path(rel_path)
-        if not os.path.exists(path):
-            raise self.ParseException(
-                    "Unable to find image file for map {0}".format(
-                        self.parse_statmap_type(contrast['statType'])))
-        return path
-
-    def expand_path(self,path):
-        if path.startswith('.'):
-            path = path.replace('.',self.ttl_relpath,1)
-        return os.path.join(self.workdir,path)
-
-    @classmethod
-    def uri_to_path(self,map_uri):
-        # NIDM namespace url, or arbitrary string just in case
-        uri = urlparse(map_uri)
-        if not uri.path or uri.scheme != 'file':
-            raise self.ParseException("Invalid location for statistic map file.")
-        return '{0}{1}'.format(uri.netloc,uri.path)
-
-    @classmethod
-    def get_unique_mapname(self,contrast):
-        type = self.parse_statmap_type(contrast['statType'])
-        return '{0} {1}'.format(contrast['contrastName'], type)
-
-    @staticmethod
-    def parse_statmap_type(stattype):
-        return urlparse(stattype).fragment or stattype
-
-    @staticmethod
-    def print_sparql_results(res):
-        for idx, row in enumerate(res.bindings):
-            rowfmt = []
-            print "Item %d" % idx
-            for key, val in sorted(row.items()):
-                rowfmt.append('%s-->%s' % (key, val.decode()))
-            print '\n'.join(rowfmt)
